Split up yeelight code into multiple files
- No functional changes
